requirements.txt
asgiref==3.7.2
beautifulsoup4==4.10.0
Django==4.2.6
django-cors-headers==3.8.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
requests==2.31.0
pint==0.17
# djangorestframework==3.14.0